Add files via upload
Analysis of VBS file from 10-31-2023. This turned out to be the same malware that I analyzed from a JS file. I was able to get a bit further in the analysis with this one. I found what looks like the final stage dll. I did not have the exe to do more in-depth analysis.